requirements.txt
PyMuPDF
pytesseract
Pillow
pandas
streamlit